requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Login==0.6.2
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
pymongo==4.5.0